+++ src/candlewick/multibody/Visualizer.cpp
@@ -200,11 +200,12 @@ void Visualizer::takeScreenshot(std::string_view filename) {
                            Uint16(height), filename);
 }
 
-void Visualizer::startRecording(std::string_view filename) {
+void Visualizer::startRecording(std::string_view filename, int fps) {
   if (m_videoRecorder.isRecording())
     terminate_with_message("Recording stream was already opened.");
 
   auto [width, height] = renderer.window.sizeInPixels();
+  m_videoSettings.fps = fps;
   m_videoRecorder.open(Uint16(width), Uint16(height), filename,
                        m_videoSettings);
   m_currentVideoFilename = filename;